api/pyproject.toml
[tool.poetry]
name = "api"
version = "0.1.0"
description = ""
authors = ["JannAnthonyBriza <jannbriza@gmail.com>"]

[tool.poetry.dependencies]
python = "~3.8"
Flask = "~1.1.1"
psycopg2-binary = "^2.8.6"


[tool.poetry.dev-dependencies]
pytest = "^4.6"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
